Do not swallow errors thrown in callback
Previously, if a synchronous error was thrown in the callback given to
toArray, the error is swallowed and never makes its way up the stack as
expected. This happens because the callback executes in a promise.then()
callback stack trace. The promise implementation catches the error and
doesn't report it other than via the "unhandledRejection" event. This is
unexpected for someone that is not otherwise using promises in their
application code.

This change uses process.nextTick to remove the callback from the
promise.then call stack.